+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 * Implement new Kong plugins.
 * Only handle plugins available on server. Starting with Kong 0.14, the list of available plugins
 can be defined in KONG_PLUGINS env var.
+* Implemented Accessible Services and Routes logic on Consumer page.
 
 ## [0.11.2](https://github.com/pantsel/konga/releases/tag/0.11.2) - 15-06-2018
 * Various Fixes.